Added Expression, SubscriptExpression, SwitchStatement, and CaseStatement classes. Updated name detection
Also updated imports in __init__.py files.

The Expression and SubscriptExpression are helper classes that
allow for a different way of creating expressions than the condition
class. They take the name of the expression function as a string, and
any number of args. The args are expected to be any python base type, or
additional Expression calls. This allows for nesting of Expressions that
are similar to how it's done in Power Automate, without adding
expressions that mirror every single expression available within
PowerAutomate.
Subscript expressions allow for expressions that use the [] operator
within Power Automate.

The SwitchStatement and CaseStatement classes export as the built-in
actions in Power Automate. They provide some error checking that Power
Automate does in order to reduce re-imports.